error_handler_worker/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-cloud-logging==3.0.0
google-cloud-pubsub==2.11.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
numpy==1.22.3
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
